Improve input handling for PICBbuild
Enhanced the functionality of PICBbuild to handle different cases for the IN.ALIGNMENTS parameter. A descriptive error message is raised if IN.ALIGNMENTS is missing either 'unique' or 'multi.primary,' causing execution to stop. A descriptive warning is issued if IN.ALIGNMENTS is missing 'multi.secondary,' but execution continues without using 'multi.secondary.'